chore: move configuration out of dawgs chore: add internal only configuration/pool generation for internal tools/tests
