+++ drivers/pg/pg.go
@@ -4,13 +4,10 @@ import (
 	"context"
 	"fmt"
 	"log/slog"
-	"time"
 
 	"github.com/jackc/pgx/v5"
-	"github.com/jackc/pgx/v5/pgxpool"
 	"github.com/specterops/dawgs"
 	"github.com/specterops/dawgs/cypher/models/pgsql"
-	"github.com/specterops/dawgs/drivers"
 	"github.com/specterops/dawgs/graph"
 )
 
@@ -20,11 +17,9 @@ const (
 	// defaultBatchWriteSize is currently set to 2k. This is meant to strike a balance between the cost of thousands
 	// of round-trips against the cost of locking tables for too long.
 	defaultBatchWriteSize = 2_000
-
-	poolInitConnectionTimeout = time.Second * 10
 )
 
-func afterPooledConnectionEstablished(ctx context.Context, conn *pgx.Conn) error {
+func AfterPooledConnectionEstablished(ctx context.Context, conn *pgx.Conn) error {
 	for _, dataType := range pgsql.CompositeTypes {
 		if definition, err := conn.LoadType(ctx, dataType.String()); err != nil {
 			if !StateObjectDoesNotExist.ErrorMatches(err) {
@@ -38,7 +33,7 @@ func afterPooledConnectionEstablished(ctx context.Context, conn *pgx.Conn) error
 	return nil
 }
 
-func afterPooledConnectionRelease(conn *pgx.Conn) bool {
+func AfterPooledConnectionRelease(conn *pgx.Conn) bool {
 	for _, dataType := range pgsql.CompositeTypes {
 		if _, hasType := conn.TypeMap().TypeForName(dataType.String()); !hasType {
 			// This connection should be destroyed since it does not contain information regarding the schema's
@@ -49,47 +44,6 @@ func afterPooledConnectionRelease(conn *pgx.Conn) bool {
 	}
 
 	return true
-}
-
-func NewPool(cfg drivers.DatabaseConfiguration) (*pgxpool.Pool, error) {
-	poolCtx, done := context.WithTimeout(context.Background(), poolInitConnectionTimeout)
-	defer done()
-
-	poolCfg, err := pgxpool.ParseConfig(cfg.PostgreSQLConnectionString())
-	if err != nil {
-		return nil, err
-	}
-
-	// TODO: Min and Max connections for the pool should be configurable
-	poolCfg.MinConns = 5
-	poolCfg.MaxConns = 50
-
-	// Bind functions to the AfterConnect and AfterRelease hooks to ensure that composite type registration occurs.
-	// Without composite type registration, the pgx connection type will not be able to marshal PG OIDs to their
-	// respective Golang structs.
-	poolCfg.AfterConnect = afterPooledConnectionEstablished
-	poolCfg.AfterRelease = afterPooledConnectionRelease
-
-	if cfg.EnableRDSIAMAuth {
-		// Only enable the BeforeConnect handler if RDS IAM Auth is enabled
-		cfg.Endpoint = cfg.LookupEndpoint()
-		poolCfg.BeforeConnect = func(ctx context.Context, connCfg *pgx.ConnConfig) error {
-			if newPoolCfg, err := pgxpool.ParseConfig(cfg.PostgreSQLConnectionString()); err != nil {
-				return err
-			} else {
-				connCfg.Password = newPoolCfg.ConnConfig.Password
-			}
-
-			return nil
-		}
-	}
-
-	pool, err := pgxpool.NewWithConfig(poolCtx, poolCfg)
-	if err != nil {
-		return nil, err
-	}
-
-	return pool, nil
 }
 
 func init() {
